Improve Widget styles getter
The new code looks at the parameters of the function in styles.js
and compute the style hash only with the props used in the parameters.
It handles the case when there is a "theme" parameter and the case with
a "props" parameter for backward compatibility. It does additional
caching for the reflexion and for the hash computation.