Improve frontend reducer and prevent frontend state leaks
The frontend reducers are automatically wrapped with shredder.
If action._type is not defined, the reducer name can be found in
action._id after the "$" (ex: workitem@111$reducer-name@222).
The Widget constructor dispatches a WIDGETS_CREATE action to set
the initial frontend state if the widget has a frontend reducer.
The frontend states are collected in componentWillUnmount.