Prevent dispatch from reducers when debugging a reducer
It seems that the exception throws only when we break in a reducer
and new dispatch are received (from the IPC for example) at the same
time. I don't know if it fixes a real potential production corner
case or if it's only a debugging behaviour.

Because the IPC events are async, it should be safe to dispatch the
new backend state from the event loop with a setTimeout.